To avoid requiring Administrator rights to install TortoiseSVN, don't use the MS merge modules for the MFC and C-Runtime libs but install them as private assemblies.
* Remove inclusion of the merge modules
* Remove custon Nant action to find the path to the merge modules
* Install the required dll's separately into their assembly folder
* On Win2k, copy the dll's from their assembly folder up into the bin folder

Since Windows2000 doesn't know about assemblies, it can't find the dll's inside their assembly folders. So we need to copy the dll's into the bin folder where the other exe files are. This of course means on Windows2000 we have those dll's twice in our installation folder. But I haven't found a way to work around that (without breaking WiX or including the files twice in the msi).

git-svn-id: https://svn.osdn.net/svnroot/tortoisesvn/trunk@5990 154be670-7969-47af-9ab0-c6fb378d2295