Start the web browser from the log dialog and the repository browser at the URL pointing to a web viewer (e.g. ViewVCS), using the new project properties webviewer:revision and webviewer:pathrevision.
The string %REVISION% in the URL is replaced by TortoiseSVN with the revision number (or "HEAD" or a revision date).
The string %PATH% in the URL is replaced by TortoiseSVN with the path relative to the repository root. For example, if you right-click in the log dialog bottom pane on a file entry /trunk/src/file and choose "View revision for path in webviewer", then the %PATH% in the url will be replaced with "/trunk/src/file".

This implements issue #201.

git-svn-id: https://svn.osdn.net/svnroot/tortoisesvn/trunk@7406 154be670-7969-47af-9ab0-c6fb378d2295