Patch from Jared Silva: The reviewer name associated with the differences generated by the comparison. If unspecified, the value defaults to the author name of the revised document or the string "Comparison" if no author information is present.
For Office XP (2002) and later, if AuthorName is not supplied for the Compare Method, deletions generated by the comparison will be under the original author name and insertions will be under the author name of the revised document.  In the general case, the original author name is the revised author name, so there is no discrepancy.  However, names can differ.  When this is the case, the differences generated by the comparison can be misleading.  The differences may suggest changes
by an author that did not actually occur.

The attached patch supplies "Comparison" as the AuthorName for Office XP (2002) and later, thus eliminating any discrepancies.  The patch also provides the added benefit of all comparisons being under the "Comparison" name.

git-svn-id: https://svn.osdn.net/svnroot/tortoisesvn/trunk@8957 154be670-7969-47af-9ab0-c6fb378d2295