Merged revision(s) 23114, 23116, 23118-23121 from trunk: First check if there's a lock owner before trying to access it. (Fixes issue #377) : file lock without owner crashes status cache
........
When w/c state is to be shown in the rev graph, we will read
the w/c info and store a number of (rev, path) pairs. If the
respective path is not versioned, we cannot use logcache-based
representations for it.

Because those nodes will never be visible in the graph (due
to the missing logcache info), it is safe to simply ignore
those pairs.

(Fixes issue #379) : Access violation when reading the w/c state for the revision graph
........
Don't crash when the author ID is NO_INDEX;
return an empty author instead.
Also, use author ID 0 for empty authors.

(Fixes issue #380) : Crash in log reporter when author information is empty
........
Follow-up to r22835: Fix the source offset handling for revprops
and merge info as well.

(Fixes issue #381) : Vector index exception when updating the revprops in log cache
........
End log iteration when the path info is not available
(i.e. invalid).

(Fixes issue #382) : Don't crash in log cache when path info is not available

........
Bring change log up-to-date.
........


git-svn-id: https://svn.osdn.net/svnroot/tortoisesvn/branches/1.7.x@23122 154be670-7969-47af-9ab0-c6fb378d2295